+++ libgadget/petaio.cpp
@@ -243,7 +243,7 @@ void
 petaio_read_snapshot(int num, const std::string OutputDir, Cosmology * CP, struct header_data * header, struct part_manager_type * PartManager, struct slots_manager_type * SlotsManager, MPI_Comm Comm)
 {
     auto fname = petaio_get_snapshot_fname(num, OutputDir);
-    int i;
+    int64_t i;
     const int ic = (num == -1);
     BigFile bf = {0};
     message(0, "Reading snapshot %s\n", fname.c_str());
@@ -354,7 +354,7 @@ petaio_read_snapshot(int num, const std::string OutputDir, Cosmology * CP, struc
         struct particle_data * parts = PartManager->Base;
         /* touch up the mass -- IC files save mass in header */
         #pragma omp parallel for
-        for(int i = 0; i < PartManager->NumPart; i++)
+        for(int64_t i = 0; i < PartManager->NumPart; i++)
         {
             if(header->MassTable[parts[i].Type] > 0)
                 parts[i].Mass = header->MassTable[parts[i].Type];
@@ -363,7 +363,7 @@ petaio_read_snapshot(int num, const std::string OutputDir, Cosmology * CP, struc
         if (!IO.UsePeculiarVelocity ) {
             /* fixing the unit of velocity from Legacy GenIC IC */
             #pragma omp parallel for
-            for(int i = 0; i < PartManager->NumPart; i++) {
+            for(int64_t i = 0; i < PartManager->NumPart; i++) {
                 /* for GenIC's Gadget-1 snapshot Unit to Gadget-2 Internal velocity unit */
                 for(int k = 0; k < 3; k++)
                     parts[i].Vel[k] *= sqrt(header->TimeSnapshot) * header->TimeSnapshot;
@@ -539,7 +539,7 @@ void petaio_alloc_buffer(BigArray * array, IOTableEntry * ent, int64_t localsize
 
 /* readout array into P struct with setters */
 void petaio_readout_buffer(BigArray * array, IOTableEntry * ent, struct conversions * conv, struct part_manager_type * PartManager, struct slots_manager_type * SlotsManager) {
-    int i;
+    int64_t i;
     /* fill the buffer */
     char * p = (char *) array->data;
     for(i = 0; i < PartManager->NumPart; i ++) {
